requirements.txt
discord.py>=2.0.0
colorama>=0.4.4
python-dotenv>=0.19.0
requests>=2.26.0
